changed: improvements after review
