--- custom-audit-examples/ror-custom-scala-serializer/src/main/scala/ScalaCustomExtendingDefaultAuditLogSerializer.scala
@@ -1,12 +1,0 @@
-import org.json.JSONObject
-import tech.beshu.ror.audit.AuditResponseContext
-import tech.beshu.ror.audit.instances.DefaultAuditLogSerializer
-
-class ScalaCustomExtendingDefaultAuditLogSerializer extends DefaultAuditLogSerializer {
-
-  override def onResponse(responseContext: AuditResponseContext): Option[JSONObject] = {
-    super.onResponse(responseContext).map { json =>
-      json.put("content", responseContext.requestContext.content)
-    }
-  }
-}
